+++ src/aws/ec2-wrapper.ts
@@ -1,29 +1,11 @@
 import { EC2 } from "aws-sdk";
-import { getAWSPagedResults, wildcardMatches } from "../utils";
-import { VPC, VPCDiscovery } from "../types";
+import { getAWSPagedResults, getValueFromTags, wildcardMatches } from "../utils";
 
 export class EC2Wrapper {
   public ec2: EC2
 
   constructor (credentials: any) {
     this.ec2 = new EC2(credentials);
-  }
-
-  /**
-   * Gets the desired vpc with the designated subnets and security groups
-   * that were set in serverless config file
-   * @returns {Promise<object>}
-   */
-  public async getVpcConfig (vpcDiscovery: VPCDiscovery): Promise<VPC> {
-    const vpc: VPC = {};
-    const vpcId = await this.getVpcId(vpcDiscovery.vpcName);
-    if (vpcDiscovery.subnetNames) {
-      vpc.subnetIds = await this.getSubnetIds(vpcId, vpcDiscovery.subnetNames);
-    }
-    if (vpcDiscovery.securityGroupNames) {
-      vpc.securityGroupIds = await this.getSecurityGroupIds(vpcId, vpcDiscovery.securityGroupNames);
-    }
-    return vpc;
   }
 
   /**
@@ -47,7 +29,7 @@ export class EC2Wrapper {
       params
     );
     if (vpcItems.length === 0) {
-      throw new Error("Invalid vpc name, it does not exist");
+      throw new Error(`VPC with tag key 'Name' and tag value '${vpcName}' does not exist.`);
     }
     return vpcItems[0].VpcId;
   }
@@ -56,17 +38,18 @@ export class EC2Wrapper {
    * Returns the promise that contains the subnet IDs
    *
    * @param {string} vpcId
-   * @param {string[]} subnetNames
+   * @param {string} tagKey
+   * @param {string[]} tagValues
    * @returns {Promise.<string[]>}
    */
-  public async getSubnetIds (vpcId: string, subnetNames: string[]): Promise<string[]> {
+  public async getSubnetIds (vpcId: string, tagKey: string, tagValues: string[]): Promise<string[]> {
     const params = {
       Filters: [{
         Name: "vpc-id",
         Values: [vpcId]
       }, {
-        Name: "tag:Name",
-        Values: subnetNames
+        Name: `tag:${tagKey}`,
+        Values: tagValues
       }]
     };
 
@@ -79,23 +62,18 @@ export class EC2Wrapper {
       params
     );
 
-    if (subnets.length === 0) {
-      throw new Error("Invalid subnet name, it does not exist");
-    }
-
-    const missingSubnetNames = subnetNames.filter((subnetName) => {
+    const missingSubnetValues = tagValues.filter((tagValue) => {
       // collect subnets by name
       const subnetsByName = subnets.filter((subnet) => {
-        const nameTag = subnet.Tags.find((tag) => tag.Key === "Name");
-        return wildcardMatches(subnetName, nameTag.Value);
+        return wildcardMatches(tagValue, getValueFromTags(subnet.Tags, tagKey));
       });
       return subnetsByName.length === 0;
     });
 
-    if (missingSubnetNames.length) {
+    if (!subnets.length || missingSubnetValues.length) {
       throw new Error(
-        `Subnets do not exist for the names: ${missingSubnetNames}. ` +
-        "Please check the names are correct or remove it."
+        `Subnets with vpc id '${vpcId}', tag key '${tagKey}' and tag values '${missingSubnetValues}' do not exist. ` +
+        "Please check the `tagKey` and `tagValues` are correct or remove it."
       );
     }
 
@@ -105,19 +83,22 @@ export class EC2Wrapper {
   /**
    *  Returns the promise that contains the security group IDs
    * @param {string} vpcId
-   * @param {string[]} securityGroupNames
+   * @param {string[]} names
+   * @param {string[]} tagKey
+   * @param {string[]} tagValues
    * @returns {Promise.<string[]>}
    */
-  public async getSecurityGroupIds (vpcId: string, securityGroupNames: string[]): Promise<string[]> {
-    const params = {
-      Filters: [{
-        Name: "vpc-id",
-        Values: [vpcId]
-      }, {
-        Name: "group-name",
-        Values: securityGroupNames
-      }]
-    };
+  public async getSecurityGroupIds (vpcId: string, names?: string[], tagKey?: string, tagValues?: string[]): Promise<string[]> {
+    // init filter by vpc id
+    const params = { Filters: [{ Name: "vpc-id", Values: [vpcId] }] };
+    // update filters with names if specified
+    if (names) {
+      params.Filters.push({ Name: "group-name", Values: names });
+    }
+    // update filters with tag and values if specified
+    if (tagKey && tagValues) {
+      params.Filters.push({ Name: `tag:${tagKey}`, Values: tagValues });
+    }
     const securityGroups = await getAWSPagedResults(
       this.ec2,
       "describeSecurityGroups",
@@ -128,22 +109,44 @@ export class EC2Wrapper {
     );
 
     if (securityGroups.length === 0) {
-      throw new Error("Invalid security group name, it does not exist");
+      const namesErrorText = names ? `, names '${names}'` : "";
+      const tagErrorText = tagKey && tagValues ? `, tag key '${tagKey}' and tag values '${tagValues}'` : "";
+      throw new Error(`Security groups with vpc id '${vpcId}'${namesErrorText}${tagErrorText} do not exist`);
     }
 
-    const missingGroupsNames = securityGroupNames.filter((groupName) => {
-      // collect security groups by name
-      const securityGroupsByName = securityGroups.filter((securityGroup) => {
-        return wildcardMatches(groupName, securityGroup.GroupName);
+    if (names) {
+      const missingGroupsNames = names.filter((groupName) => {
+        // collect security groups by name
+        const securityGroupsByName = securityGroups.filter((securityGroup) => {
+          return wildcardMatches(groupName, securityGroup.GroupName);
+        });
+        return securityGroupsByName.length === 0;
       });
-      return securityGroupsByName.length === 0;
-    });
 
-    if (missingGroupsNames.length) {
-      throw new Error(
-        `Security groups do not exist for the names: ${missingGroupsNames}. ` +
-        "Please check the names are correct or remove it."
-      );
+      if (missingGroupsNames.length) {
+        throw new Error(
+          `Security groups do not exist for the names: ${missingGroupsNames}. ` +
+          "Please check the 'names' are correct or remove it."
+        );
+      }
+    }
+
+    if (tagKey && tagValues) {
+      tagValues = tagValues || [];
+      const missingGroupsTagNames = tagValues.filter((tagValue) => {
+        // collect subnets by name
+        const groupsByName = securityGroups.filter((securityGroup) => {
+          const groupTagValue = getValueFromTags(securityGroup.Tags, tagKey);
+          return groupTagValue === tagValue;
+        });
+        return groupsByName.length === 0;
+      });
+      if (missingGroupsTagNames.length) {
+        throw new Error(
+          `Security groups do not exist for the tag '${tagKey}' and tag values: '${missingGroupsTagNames}'. ` +
+          "Please check the 'tagKey' and 'tagValues' are correct or remove it."
+        );
+      }
     }
 
     return securityGroups.map((group) => group.GroupId);