+++ src/index.ts
@@ -1,8 +1,9 @@
 "use strict";
 
-import { ServerlessInstance } from "./types";
+import { FuncVPCDiscovery, ServerlessInstance, VPCDiscovery } from "./types";
 import { LambdaFunction } from "./common/lambda-function";
 import Globals from "./globals";
+import { validateVPCDiscoveryConfig } from "./validation";
 
 class VPCPlugin {
   private serverless: ServerlessInstance;
@@ -35,21 +36,50 @@ class VPCPlugin {
    */
   public validateCustomVPCDiscoveryConfig (): void {
     const config = this.serverless.service.custom;
-    if (config && !config.vpcDiscovery && config.vpc) {
-      config.vpcDiscovery = config.vpc;
-      Globals.logWarning(
-        "The `vpc` option of `custom` config is deprecated and will be removed in the future. " +
-        "Please use the `vpcDiscovery` option instead."
-      );
+    const vpcDiscovery = config && config.vpcDiscovery;
+
+    if (vpcDiscovery) {
+      // support backward compatibility
+      this.updateVPCDiscoveryConfigCompatibility(vpcDiscovery);
+
+      // validate config
+      try {
+        // the validateVPCDiscoveryConfig is general for custom and func configs
+        // so try catch for extend error message with `custom.vpcDiscovery` as a source
+        validateVPCDiscoveryConfig(vpcDiscovery);
+      } catch (e) {
+        throw new Error(
+          `The \`custom.vpcDiscovery\` is not configured correctly: \n${e} ` + " Please see README for proper setup."
+        );
+      }
     }
-    const vpc = config && config.vpcDiscovery;
-    // Checking the vpcDiscovery config is setup correctly if exists
-    if (vpc && (vpc.vpcName == null || (vpc.subnetNames == null && vpc.securityGroupNames == null))) {
-      throw new Error(
-        "The `custom.vpcDiscovery` is not configured correctly. " +
-        "You must specify the vpcName and at least one of subnetNames or securityGroupNames. " +
-        "Please see README for proper setup."
-      );
+  }
+
+  /**
+   * This function update VPC Discovery config to support version 2.x config structure and show warning errors
+   */
+  public updateVPCDiscoveryConfigCompatibility (vpcDiscovery: VPCDiscovery | FuncVPCDiscovery): void {
+    // support backward compatibility
+    if (vpcDiscovery && (vpcDiscovery.subnetNames || vpcDiscovery.securityGroupNames)) {
+      // convert `vpcDiscovery.subnetNames` or `vpcDiscovery.securityGroupNames` to the new config structure
+      if (!vpcDiscovery.subnets && !vpcDiscovery.securityGroups) {
+        if (vpcDiscovery.subnetNames) {
+          vpcDiscovery.subnets = [{ tagKey: "Name", tagValues: vpcDiscovery.subnetNames }];
+        }
+        if (vpcDiscovery.securityGroupNames) {
+          vpcDiscovery.securityGroups = [{ names: vpcDiscovery.securityGroupNames }];
+        }
+        Globals.logWarning(
+          "The `vpcDiscovery.subnetNames` and `vpcDiscovery.securityGroupNames` options are deprecated " +
+          "and will be removed in the future. Please see README for proper setup."
+        );
+      } else {
+        // log warning in case mixed config are specified
+        Globals.logWarning(
+          "The `vpcDiscovery.subnetNames` and `vpcDiscovery.securityGroupNames` are deprecated " +
+          "and will not be applied. Please remove mentioned option to not see this warning message."
+        );
+      }
     }
   }
 
@@ -71,46 +101,44 @@ class VPCPlugin {
   public async updateFunctionsVpcConfig (): Promise<object> {
     Globals.logInfo("Updating VPC config...");
     const service = this.serverless.service;
+    const functions = service.functions || {};
 
     // Sets the serverless's vpc config
-    if (service.functions) {
-      // loop through the functions and update VPC config
-      // eslint-disable-next-line guard-for-in
-      for (const funcName in service.functions) {
-        // eslint-disable-next-line
-        const func = service.functions[funcName];
-        const funcVPC = await this.lambdaFunction.getFuncVPC(funcName, func.vpcDiscovery);
+    // loop through the functions and update VPC config
+    for (const funcName of Object.keys(functions)) {
+      const func = service.functions[funcName];
+      this.updateVPCDiscoveryConfigCompatibility(func.vpcDiscovery);
+      const funcVPC = await this.lambdaFunction.getFuncVPC(funcName, func.vpcDiscovery);
 
-        if (!funcVPC) {
-          continue;
-        }
+      if (!funcVPC) {
+        continue;
+      }
 
-        // init vpc empty config in case not exists
-        func.vpc = func.vpc || {};
-        // log warning in case vpc.subnetIds and vpcDiscovery.subnetNames are specified.
-        if (func.vpc.subnetIds && func.vpcDiscovery && func.vpcDiscovery.subnetNames) {
-          Globals.logWarning(
-            `vpc.subnetIds' are specified for the function '${funcName}' 
-            and overrides 'vpcDiscovery.subnetNames' discovery config.`
-          );
-        }
-        // log warning in case vpc.securityGroupIds and vpcDiscovery.securityGroupNames are specified.
-        if (func.vpc.securityGroupIds && func.vpcDiscovery && func.vpcDiscovery.securityGroupNames) {
-          Globals.logWarning(
-            `vpc.securityGroupIds' are specified for the function '${funcName}' 
-            and overrides 'vpcDiscovery.securityGroupNames' discovery config.`
-          );
-        }
+      // init vpc empty config in case not exists
+      func.vpc = func.vpc || {};
+      // log warning in case vpc.subnetIds and vpcDiscovery.subnetNames are specified.
+      if (func.vpc.subnetIds && func.vpcDiscovery && func.vpcDiscovery.subnets) {
+        Globals.logWarning(
+          `vpc.subnetIds' are specified for the function '${funcName}' ` +
+          "and overrides 'vpcDiscovery.subnets' discovery config."
+        );
+      }
+      // log warning in case vpc.securityGroupIds and vpcDiscovery.securityGroupNames are specified.
+      if (func.vpc.securityGroupIds && func.vpcDiscovery && func.vpcDiscovery.securityGroups) {
+        Globals.logWarning(
+          `vpc.securityGroupIds' are specified for the function '${funcName}' ` +
+          "and overrides 'vpcDiscovery.securityGroups' discovery config."
+        );
+      }
 
-        // set vpc.subnetIds if it does not exists
-        if (!func.vpc.subnetIds && funcVPC.subnetIds) {
-          func.vpc.subnetIds = funcVPC.subnetIds;
-        }
+      // set vpc.subnetIds if it does not exists
+      if (!func.vpc.subnetIds && funcVPC.subnetIds) {
+        func.vpc.subnetIds = funcVPC.subnetIds;
+      }
 
-        // set vpc.securityGroupIds if it does not exists
-        if (!func.vpc.securityGroupIds && funcVPC.securityGroupIds) {
-          func.vpc.securityGroupIds = funcVPC.securityGroupIds;
-        }
+      // set vpc.securityGroupIds if it does not exists
+      if (!func.vpc.securityGroupIds && funcVPC.securityGroupIds) {
+        func.vpc.securityGroupIds = funcVPC.securityGroupIds;
       }
     }